Broke video video into separate readable scripts
